+++ ChangeLog.md
@@ -8,6 +8,7 @@
 * Rewrite SOP class extended negotiation support (#740)
 * Bug fix: extended negotiations have been accepted by default, but should be explicitly accepted (#900)
 * Bug fix: Send only PrintManagementMetaSopClassUid on AssociationRequest instead of all the various SopClassUids of the NCreate and NSet requests (#667)
+* Global static property DicomValidation.PerformValidation to turn off validation for every DicomDataset. 
 
 #### v.4.0.2 (7/30/2019)
 * Bug fix: prevent resource leak when DesktopNetworkListener waits for new TCP clients